# backend/api.yaml
openapi: 3.0.0
info:
  title: TastyBites API Documentation
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
  version: 1.0 
servers:
  - url: http://localhost:5000/
    description:  Local server 
  - url: https://prod.com/
    description:  Pre Production server
  - url: https://test.com/
    description:  Production server
components:
  securitySchemes:
    ApiTokenss:        # arbitrary name for the security scheme
          
      type: http
      scheme: bearer
    
    ApiKey:        # arbitrary name for the security scheme
      type: apiKey
      in: header       # can be "header", "query" or "cookie"
      name: apikey
paths:
...
  /api/restaurantOrderData: 
    post:
      tags:
          - Food Item
      summary: to display order by restaurant id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                RestaurantId:
                  type: string
      responses: 
        '200':
          description: OK
        '400':
          description: Error Bad Request
        '404':
          description: Server Error
        default:
          description: Unexpected error
...
  /api/updateOrderConfirmation: 
    put:
      tags:
          - Restaurant
      summary: to confirm the order by the restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: string
                foodId:
                  type: string
      responses: 
        '200':
          description: OK
        '400':
          description: Error Bad Request
        '404':
          description: Server Error
        default:
          description: Unexpected error